Add icon placement policy, icon drag, and cycle menu
Implemented the three planned features: configurable icon placement
policy (`bottom`/`top`/`left`/`right`/`smart`) stored in `config.json`
and applied via `place_icon` replacing the old static slot counter; icon
window drag using the existing `s_drag` infrastructure with a new
`drag_window` field so that only the icon is moved and its position
persisted in `client->icon_x`/`y` for re-use on re-iconification; and
a keyboard cycle menu popup (Openbox-like) for both window cycling and
icon cycling, replacing the direct focus/restore dispatch with
a floating list window navigable via `Up`/`Down`/`Enter`/`Esc` and
clickable with the mouse.

+++ src/render/desktop.c
@@ -312,7 +312,7 @@ int desktop_render_clients(desktop_td *desktop, bool is_current)
             right = (uint16_t) client->layout.frame_extents.right;
             top = (uint16_t) client->layout.frame_extents.top;
             bottom = (uint16_t) client->layout.frame_extents.bottom;
-            title_h = client->title_height;
+            title_h = (uint16_t) client->title_height;
         inner_w = (client->layout.geometry.cur.dim.w > left + right)
             ? (uint16_t) (client->layout.geometry.cur.dim.w - left - right)
             : 1;
